chore: rename components as per file naming convention
